# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Delphi API",
    "description": "Compare organizational documents with traceable evidence and human review.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/runs/{run_id}/findings": {
      "get": {
        "tags": [
          "findings"
        ],
        "summary": "List Findings",
        "operationId": "list_findings",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/FindingResponse"
                  },
                  "title": "Response List Findings"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Request Entity Too Large"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal Server Error"
          },
          "502": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad Gateway"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Service Unavailable"
          },
          "504": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Gateway Timeout"
          }
        }
      }
    },
    "/api/findings/{finding_id}/evidence": {
      "get": {
        "tags": [
          "findings"
        ],
        "summary": "Get Finding Evidence",
        "operationId": "get_finding_evidence",
        "parameters": [
          {
            "name": "finding_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Finding Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/EvidenceResponse"
                  },
                  "title": "Response Get Finding Evidence"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Request Entity Too Large"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal Server Error"
          },
          "502": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad Gateway"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Service Unavailable"
          },
          "504": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Gateway Timeout"
          }
        }
      }
    },
    "/api/findings/{finding_id}/review": {
      "put": {
        "tags": [
          "findings"
        ],
        "summary": "Update Finding Review",
        "operationId": "update_finding_review",
        "parameters": [
          {
            "name": "finding_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Finding Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateReview"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewUpdated"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Request Entity Too Large"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal Server Error"
          },
          "502": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad Gateway"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Service Unavailable"
          },
          "504": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Gateway Timeout"
          }
        }
      }
    },
    "/api/sources/{source_id}": {
      "get": {
        "tags": [
          "findings"
        ],
        "summary": "Get Source",
        "operationId": "get_source",
        "parameters": [
          {
            "name": "source_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Source Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SourceResponse"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Request Entity Too Large"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal Server Error"
          },
          "502": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad Gateway"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Service Unavailable"
          },
          "504": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Gateway Timeout"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ErrorDetail": {
        "properties": {
          "location": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Type"
          }
        },
        "type": "object",
        "required": [
          "location",
          "message",
          "type"
        ],
        "title": "ErrorDetail"
      },
      "ErrorResponse": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "details": {
            "items": {
              "$ref": "#/components/schemas/ErrorDetail"
            },
            "type": "array",
            "title": "Details"
          }
        },
        "type": "object",
        "required": [
          "code",
          "message",
          "details"
        ],
        "title": "ErrorResponse"
      },
      "EvidenceResponse": {
        "properties": {
          "source_id": {
            "type": "string",
            "format": "uuid",
            "title": "Source Id"
          },
          "document_id": {
            "type": "string",
            "format": "uuid",
            "title": "Document Id"
          },
          "filename": {
            "type": "string",
            "title": "Filename"
          },
          "side": {
            "type": "string",
            "enum": [
              "before",
              "after"
            ],
            "title": "Side"
          },
          "clause_no": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Clause No"
          },
          "original_text": {
            "type": "string",
            "title": "Original Text"
          },
          "excerpt": {
            "type": "string",
            "title": "Excerpt"
          },
          "locator": {
            "additionalProperties": true,
            "type": "object",
            "title": "Locator"
          },
          "evidence_role": {
            "type": "string",
            "enum": [
              "before",
              "after",
              "context"
            ],
            "title": "Evidence Role"
          },
          "start_offset": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Start Offset"
          },
          "end_offset": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "End Offset"
          }
        },
        "type": "object",
        "required": [
          "source_id",
          "document_id",
          "filename",
          "side",
          "clause_no",
          "original_text",
          "excerpt",
          "locator",
          "evidence_role",
          "start_offset",
          "end_offset"
        ],
        "title": "EvidenceResponse"
      },
      "FindingResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "run_id": {
            "type": "string",
            "format": "uuid",
            "title": "Run Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "change_type": {
            "type": "string",
            "enum": [
              "retained",
              "reworded",
              "transferred",
              "split",
              "merged",
              "new",
              "potentially_missing",
              "changed",
              "structure_changed"
            ],
            "title": "Change Type"
          },
          "issue_type": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "overlap",
                  "potential_conflict",
                  "modality_changed",
                  "scope_changed",
                  "insufficient_evidence"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Issue Type"
          },
          "before_function_ids": {
            "items": {
              "type": "string",
              "format": "uuid"
            },
            "type": "array",
            "title": "Before Function Ids"
          },
          "after_function_ids": {
            "items": {
              "type": "string",
              "format": "uuid"
            },
            "type": "array",
            "title": "After Function Ids"
          },
          "explanation": {
            "type": "string",
            "title": "Explanation"
          },
          "recommendation": {
            "type": "string",
            "title": "Recommendation"
          },
          "source_ids": {
            "items": {
              "type": "string",
              "format": "uuid"
            },
            "type": "array",
            "title": "Source Ids"
          },
          "search": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SearchCoverage"
              },
              {
                "type": "null"
              }
            ]
          },
          "review": {
            "$ref": "#/components/schemas/ReviewResponse"
          }
        },
        "type": "object",
        "required": [
          "id",
          "run_id",
          "title",
          "change_type",
          "issue_type",
          "before_function_ids",
          "after_function_ids",
          "explanation",
          "recommendation",
          "source_ids",
          "search",
          "review"
        ],
        "title": "FindingResponse"
      },
...
      "ReviewResponse": {
        "properties": {
          "finding_id": {
            "type": "string",
            "format": "uuid",
            "title": "Finding Id"
          },
          "status": {
            "type": "string",
            "enum": [
              "unreviewed",
              "confirmed",
              "needs_clarification",
              "rejected"
            ],
            "title": "Status"
          },
          "note": {
            "type": "string",
            "title": "Note"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "finding_id",
          "status",
          "note",
          "updated_at"
        ],
        "title": "ReviewResponse"
      },
      "ReviewUpdated": {
        "properties": {
          "finding_id": {
            "type": "string",
            "format": "uuid",
            "title": "Finding Id"
          },
          "status": {
            "type": "string",
            "enum": [
              "unreviewed",
              "confirmed",
              "needs_clarification",
              "rejected"
            ],
            "title": "Status"
          },
          "note": {
            "type": "string",
            "title": "Note"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "review_revision": {
            "type": "integer",
            "title": "Review Revision"
          }
        },
        "type": "object",
        "required": [
          "finding_id",
          "status",
          "note",
          "updated_at",
          "review_revision"
        ],
        "title": "ReviewUpdated"
      },
...
      "SearchCoverage": {
        "properties": {
          "method": {
            "type": "string",
            "const": "semantic_all_after_batches",
            "title": "Method"
          },
          "complete": {
            "type": "boolean",
            "title": "Complete"
          },
          "reviewed_source_ids": {
            "items": {
              "type": "string",
              "format": "uuid"
            },
            "type": "array",
            "title": "Reviewed Source Ids"
          },
          "candidate_source_ids": {
            "items": {
              "type": "string",
              "format": "uuid"
            },
            "type": "array",
            "title": "Candidate Source Ids"
          },
          "errors": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Errors"
          },
          "input_partial": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Input Partial"
          }
        },
        "type": "object",
        "required": [
          "method",
          "complete",
          "reviewed_source_ids",
          "candidate_source_ids",
          "errors"
        ],
        "title": "SearchCoverage"
      },
      "SourceResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "document_id": {
            "type": "string",
            "format": "uuid",
            "title": "Document Id"
          },
          "clause_no": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Clause No"
          },
          "parent_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parent Id"
          },
          "original_text": {
            "type": "string",
            "title": "Original Text"
          },
          "locator": {
            "additionalProperties": true,
            "type": "object",
            "title": "Locator"
          }
        },
        "type": "object",
        "required": [
          "id",
          "document_id",
          "clause_no",
          "parent_id",
          "original_text",
          "locator"
        ],
        "title": "SourceResponse"
      },
...
      "UpdateReview": {
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "unreviewed",
              "confirmed",
              "needs_clarification",
              "rejected"
            ],
            "title": "Status"
          },
          "note": {
            "type": "string",
            "maxLength": 4000,
            "title": "Note"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "status",
          "note"
        ],
        "title": "UpdateReview"
      }
    }
  }
}